slightly more cleanup
