+++ src/lib.rs
@@ -2,3 +2,4 @@ pub mod common;
 pub mod day1;
 pub mod day2;
 pub mod day3;
+pub mod vec2d;
